refactor emsfrontendbundle
