+++ EMSFrontendBundle/Resources/config/services.xml
@@ -5,14 +5,14 @@
         http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-       <parameter key="emsch.request_listener.class">EMS\ClientHelperBundle\Frontend\EventListener\RequestListener</parameter>
+       <parameter key="emsch.request_listener.class">EMS\ClientHelperBundle\EMSFrontendBundle\EventListener\RequestListener</parameter>
        
-       <parameter key="emsch.twig.extension.helper.class">EMS\ClientHelperBundle\Frontend\Twig\HelperExtension</parameter>
-       <parameter key="emsch.twig.extension.trans.class">EMS\ClientHelperBundle\Frontend\Twig\TranslationExtension</parameter>
+       <parameter key="emsch.twig.extension.helper.class">EMS\ClientHelperBundle\EMSFrontendBundle\Twig\HelperExtension</parameter>
+       <parameter key="emsch.twig.extension.trans.class">EMS\ClientHelperBundle\EMSFrontendBundle\Twig\TranslationExtension</parameter>
        
-       <parameter key="emsch.request.service.class">EMS\ClientHelperBundle\Frontend\Service\RequestService</parameter>
-       <parameter key="emsch.client_request.class">EMS\ClientHelperBundle\Frontend\Elasticsearch\ClientRequest</parameter>
-       <parameter key="emsch.client_request_factory.class">EMS\ClientHelperBundle\Frontend\Elasticsearch\ClientRequestFactory</parameter>
+       <parameter key="emsch.request.service.class">EMS\ClientHelperBundle\EMSFrontendBundle\Service\RequestService</parameter>
+       <parameter key="emsch.client_request.class">EMS\ClientHelperBundle\EMSFrontendBundle\Elasticsearch\ClientRequest</parameter>
+       <parameter key="emsch.client_request_factory.class">EMS\ClientHelperBundle\EMSFrontendBundle\Elasticsearch\ClientRequestFactory</parameter>
     </parameters>
     
     <services>
