+++ EMSFrontendBundle/Tests/Controller/DefaultControllerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace EMS\ClientHelperBundle\Frontend\Tests\Controller;
+namespace EMS\ClientHelperBundle\EMSFrontendBundle\Tests\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
